+++ src/c#/main/screens/ScreenType.cs
@@ -1,0 +1,7 @@
+namespace beyondnations {
+
+    public enum ScreenType {
+        TITLE,
+        WORLD
+    }
+}
