+++ src/c#/main/screens/ScreenType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0963ad0b7e012354984db52e5d80df95
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
