+++ src/c#/main/screens/WorldScreen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 991eba815eb09684d8428e76f8ad2554
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
